[fix] handler fn names & [add] feat to access height and width of component: failed
The Height and width of component size is necessary in order to create similar sized container upon add entry loading/success conditional component renders. If the size of the original form render is obtained, Similar sized components can be renderd when loading and success states are rendered. The intention is to keep the container the same size as to not break the flow of the page. Current attempt has failed. A ref of the component AddEntryForm, has not been created and the dimensions cannot be obtained. More research into useRef will be necessary