[feat] add date picker component & [feat] begin creation of selecting proper entries to display
Day component should display entries based on a particular date. This selecting process must happen either within this component itself or within the react-query hook. The notes placed in the useEffect hook are an implementation that selects from entries prop and filters them directly within the component. It should be noted that the notes only apply to an effect that should run on initial mount. It is very likely that this implementation will be scrapped.