[fix] create new date object for incoming arg
Incoming argument 'date' was a state variable. It was being mutated in place and causing the original state value to change as well. Big no-no for multiple reasons: firstly, state vars should only change with set State and second the original date value was not meant to be altered from within this particular function.